+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkRuleRedeclarationInspection.kt
@@ -5,6 +5,7 @@ import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemHighlightType.GENERIC_ERROR
 import com.intellij.codeInspection.ProblemHighlightType.WEAK_WARNING
 import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
 import com.jetbrains.python.psi.PyStatement
 import com.jetbrains.snakecharm.SnakemakeBundle
@@ -12,13 +13,14 @@ import com.jetbrains.snakecharm.inspections.quickfix.RenameElementWithoutUsagesQ
 import com.jetbrains.snakecharm.lang.psi.*
 import com.jetbrains.snakecharm.lang.psi.stubs.SmkCheckpointNameIndex
 import com.jetbrains.snakecharm.lang.psi.stubs.SmkRuleNameIndex
+import com.jetbrains.snakecharm.lang.psi.stubs.SmkUseNameIndex
 import com.jetbrains.snakecharm.lang.psi.types.AbstractSmkRuleOrCheckpointType
 
 class SmkRuleRedeclarationInspection : SnakemakeInspection() {
     override fun buildVisitor(
-            holder: ProblemsHolder,
-            isOnTheFly: Boolean,
-            session: LocalInspectionToolSession
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+        session: LocalInspectionToolSession
     ) = object : SnakemakeInspectionVisitor(holder, session) {
         private val localRules by lazy {
             holder.file.let { psiFile ->
@@ -40,6 +42,16 @@ class SmkRuleRedeclarationInspection : SnakemakeInspection() {
                 }
             }
         }
+        private val localUses by lazy {
+            holder.file.let { psiFile ->
+                when (psiFile) {
+                    is SmkFile -> {
+                        psiFile.collectUses().map { it.second }
+                    }
+                    else -> emptyList()
+                }
+            }
+        }
 
         override fun visitSmkRule(rule: SmkRule) {
             visitSMKRuleLike(rule)
@@ -47,6 +59,10 @@ class SmkRuleRedeclarationInspection : SnakemakeInspection() {
 
         override fun visitSmkCheckPoint(checkPoint: SmkCheckPoint) {
             visitSMKRuleLike(checkPoint)
+        }
+
+        override fun visitSmkUse(use: SmkUse) {
+            visitSMKRuleLike(use)
         }
 
         private fun visitSMKRuleLike(ruleLike: SmkRuleLike<SmkRuleOrCheckpointArgsSection>) {
@@ -61,11 +77,17 @@ class SmkRuleRedeclarationInspection : SnakemakeInspection() {
                 ruleLike, nameToCheck, SmkCheckpointNameIndex.KEY, SmkCheckPoint::class.java
             ) { localCheckpoints }
 
-            val resolveResults = ruleResolveResults + cpResolveResults
+            val usesResolveResults: Collection<PsiElement> =
+                AbstractSmkRuleOrCheckpointType.findAvailableRuleLikeElementByName(
+                    ruleLike, nameToCheck, SmkUseNameIndex.KEY, SmkUse::class.java
+                ) { localUses }
+
+            val resolveResults = ruleResolveResults + cpResolveResults + usesResolveResults
 
             if (resolveResults.isEmpty()) {
                 return
             }
+
             if (resolveResults.size == 1 && resolveResults.single() == ruleLike) {
                 return
             }
@@ -98,7 +120,7 @@ class SmkRuleRedeclarationInspection : SnakemakeInspection() {
         }
 
         private fun registerProblem(
-            ruleLike: SmkRuleLike<SmkRuleOrCheckpointArgsSection>,
+            ruleLike: SmkSection,
             msg: String,
             severity: ProblemHighlightType
         ) {
@@ -114,5 +136,6 @@ class SmkRuleRedeclarationInspection : SnakemakeInspection() {
             )
         }
     }
+
     override fun getDisplayName(): String = SnakemakeBundle.message("INSP.NAME.rule.redeclaration")
 }