+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/stubs/SmkStubs.kt
@@ -1,10 +1,10 @@
 package com.jetbrains.snakecharm.lang.psi.stubs
 
 import com.intellij.psi.stubs.NamedStub
-import com.jetbrains.snakecharm.lang.psi.SmkCheckPoint
-import com.jetbrains.snakecharm.lang.psi.SmkRule
-import com.jetbrains.snakecharm.lang.psi.SmkSubworkflow
+import com.jetbrains.snakecharm.lang.psi.*
 
-interface SmkRuleStub: NamedStub<SmkRule>
-interface SmkCheckpointStub: NamedStub<SmkCheckPoint>
-interface SmkSubworkflowStub: NamedStub<SmkSubworkflow>
+interface SmkRuleStub : NamedStub<SmkRule>
+interface SmkCheckpointStub : NamedStub<SmkCheckPoint>
+interface SmkSubworkflowStub : NamedStub<SmkSubworkflow>
+interface SmkModuleStub : NamedStub<SmkModule>
+interface SmkUseStub : NamedStub<SmkUse>